[Feature]: Allow listing available key resources such as gpu, region, and backends #2142 (WIP)
Moved the /gpus API to a separate routing; also moved GPU-related logic from runs.
